Move Elastic config into seperate object
